Rename coroutine to coro
