+++ include/yaclib/coro/detail/on_awaiter.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <yaclib/coroutine/coroutine.hpp>
-#include <yaclib/coroutine/detail/promise_type.hpp>
+#include <yaclib/coro/coroutine.hpp>
+#include <yaclib/coro/detail/promise_type.hpp>
 #include <yaclib/exe/executor.hpp>
 #include <yaclib/exe/job.hpp>
