+++ include/yaclib/coro/detail/promise_type.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <yaclib/async/detail/result_core.hpp>
-#include <yaclib/coroutine/coroutine.hpp>
+#include <yaclib/coro/coroutine.hpp>
 #include <yaclib/util/detail/atomic_counter.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
