+++ src/CMakeLists.txt
@@ -26,7 +26,7 @@ add_subdirectory(fault)
 add_subdirectory(std)
 
 if (YACLIB_CORO)
-  add_subdirectory(coroutine)
+  add_subdirectory(coro)
 endif ()
 
 configure_file(config.hpp.in ${YACLIB_BINARY_DIR}/include/yaclib/config.hpp)
