+++ src/coro/CMakeLists.txt
@@ -1,14 +1,14 @@
 list(APPEND YACLIB_INCLUDES
-  ${YACLIB_INCLUDE_DIR}/coroutine/await.hpp
-  ${YACLIB_INCLUDE_DIR}/coroutine/coroutine.hpp
-  ${YACLIB_INCLUDE_DIR}/coroutine/future_traits.hpp
-  ${YACLIB_INCLUDE_DIR}/coroutine/on.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/await.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/coroutine.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/future_traits.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/on.hpp
   )
 
 list(APPEND YACLIB_HEADERS
-  ${YACLIB_INCLUDE_DIR}/coroutine/detail/await_awaiter.hpp
-  ${YACLIB_INCLUDE_DIR}/coroutine/detail/promise_type.hpp
-  ${YACLIB_INCLUDE_DIR}/coroutine/detail/on_awaiter.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/detail/await_awaiter.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/detail/promise_type.hpp
+  ${YACLIB_INCLUDE_DIR}/coro/detail/on_awaiter.hpp
   )
 
 list(APPEND YACLIB_SOURCES
